tests: add means to install released binaries
This commit introduces the RedpandaInstaller as a means to manage and
install binaries across nodes in a RedpandaService. It allows developers
to, before starting Redpanda, install and run with a different set of
binaries, maintaining the old binaries for the sake of upgrade or a
post-test revert.

The underlying mechanism used to install binaries is the downloading of
tarballs from the Redpanda public repository. Downloads are placed in a
specific directory, and /opt/redpanda is maintained as a symlink to the
currently active version.

Upon starting up a test, some sanity checks are performed to ensure
the versions across different nodes match. Upon shutting down a test,
cleanup is invoked to delete all downloads and restore the original
binary.