tests: add means to fetch previous feature release
In thinking about continuously testing upgrades, a critical upgrade path
is that from the highest feature release below the current branch.

This commit adds this capability to the RedpandaInstaller by traversing
a list of releases fetched from Github for a valid candidate.

A test is added to demonstrate this, though verification of the versions
is somewhat limited because the previous feature branch today (v21.11.x)
doesn't have a great way to get the version from a running Redpanda
cluster.