c/topics_frontend: check if update is in progress when finishing it
When finishing update we may reduce number of records replicated to
raft_0 log by simply checking if there is in progress update for
requested `ntp` and if not we may just return early without the need of
replicating `finish_update` command. This approach is safe as updates
are finished as a part of reconciliation loop and will be retried if
failed.

Signed-off-by: Michal Maslanka <michal@redpanda.com>