archival: Implement manifest truncation in archival
Adds method 'maybe_truncate_manifest' method to ntp_archiver. This
method verifies content of the bucket by using HeadObject request. If
the head of the log can't be found in the bucket the manifest is
truncated. The archival metadata STM is also truncated and new manifest
is re-uploaded to S3.