cloud_storage: Don't cache start offset
The starting offset is cached in the remote_partition. This may cause
problems when the manifest is truncated. This commit removes the caching
mechanism. The performance impact is not critical since the start offset
is available in manifest with O(1) access time.

(cherry picked from commit 1dfed1624619de9646e8a9046d55afa4e2b12117)