pandaproxy: Add latency metric to public endpoint
This patch adds request latency metrics for pandaproxy and the schema
registry.

* repdanda_rest_proxy_request_latency_seconds
* redpanda_schema_registry_request_latency_seconds
    * Description: latency of requests processed by the respective
      server which includes the time spent waiting for resources to
      become available
    * Labels: shard, operation
    * Aggregation: shard, operation

This patch updates the server constructor to accept the name of
the metrics group under which to publish metrics for the new endpoint.
The server name, which is already an argument to the constructor, cannot
be used for this because we wish to publish 'pandaproxy' metrics under
the 'rest_proxy' name.