net: Ensure server catches all thrown exceptions
- The method `is_disconnect_exception` throws an exception that may be
of any type, and only handles a few cases. In the case the exception is
not handled, the method returns std::nullopt, so the caller can
determine that the exception type "was not of a disconnect".

- However there may be other cases where unexpected exceptions may be
thrown, this patch sets up a catch handler for these cases so that
redpanda never ignores any exceptional futures.

- This fixes a bug for an unhandled exception that was observed when rpk
was sending erroneously formatted messages to redpanda, causing
`parse_header` to throw after observing a negative length size.