Pass node report by reference
We currently pass the node_report by value which leads to a copy which
may be large if the report is large (especially if there are many
partitions). There doesn't seem to be any reason we can't pass this
by value, as we are co_awaiting the return of the invoke_all.

This was implicated in the OOM in #5137. There were too parts to that
OOM: first the lambda was very large due to the issue fixed here and
second seastar was copying the lambda unnecessarily.

Fixes #5137.