cmake: adds option to enable distcc
adds a variable ENABLE_DISTCC which is off by default. If enabled,
distcc is either appended to the compiler launcher, or set alone
depending on what value was given to it earlier.

As ccache.cmake clobbers the compiler launcher, we include distcc after
that so the compiler launcher contains "ccache;distcc" if both are
found.