ducky: add read-write verifier
rw_verifier is a replacement for verified producer and
consumer services. rw_verifier is a workload generator
and an online linear and incremental consistency checker
capable of running for hours without an overhead.

The validation happens as the write / read go. No need
to pass written/read data to the test over ssh for vali-
dation.

The app exposes control and metrics over http. So there
is no need to eyeball the number of records:
 - start the workload
 - use the api to check that the app doing a progress at
   key points of the test
 - stop the app when the need without waiting for the
   remaining messages to pass