tests: http imposter as library
multiple tests use http imposter and since tests are not built as a
library they repeat the definition of the http imposter server, in s3 as
well as the archival libs the s3_imposter_fixture is defined.

this set of changes pulls http imposter into a top level lib so the code
requiring a mock http server can link against this library.

also a fluent api is added to register mock urls, which can be used in
test body using fixture as:

    when()
      .request("/gettoken")
      .with_method(ss::httpd::POST)
      .then_reply_with("token");