cluster: give local_monitor its own tick loop
This decouples it from relying on health_monitor_backend::tick
to keep it up to date.

Also initialize it up-front during ::start.

This removes the explicit call to update_state from
health_monitor_backend when it prepares a health report:
the disk info in health report will now only be as recent
as the last tick.

We can also much increase the frequency with which we read
the local disk utilization, as it is no longer coupled to
the frequency with which we gather all the bulky node
healthy structures (partitions lists etc), which enables
systems that depend on disk space knowledge to respond more quickly
to changes.