s3: uses credential appliers for s3 client
s3 client pool uses a credential applier, which is responsible for
injecting authentication headers to each outgoing http request. the
applier could contain gcp or aws based credentials and applies the
corresponding headers to outgoing requests.

this change also refactors the signature object to reside in the
cloud_roles namespace. the hierarchy of namespaces is such that s3
depends on cloud_roles, so we need to make sure that cloud_roles
contains all credential related objects and does not depend on s3 to
create a cyclic dependency.