c/members_backend: store revision of previous node decommissioning
In order to cancel only those partition movements which were requested
by node decommissioning we need to know which partition movements were
requested before the node was decommissioned.

In order to calculate the boundary and be able to cancel only
decommission related partition movements when recommissioning a node we
store revision of last node decommission request.

All the controller commands, both partition movements and node
operations can be sequenced using controller log offset. The offset
allow us to set a boundary between the moves which were scheduled before
node was decommissioned and ones which are a result of decommissioning.

Signed-off-by: Michal Maslanka <michal@redpanda.com>