tests: clarify comment to save executable
If the 'CI' environment variable is unset, we save the executable by
default. The comment around not saving the executable was a bit
confusing, since it expects this to only be the case when running on a
developer workstation. This isn't the case, e.g. if running `ducktape`
against a remote cluster without setting the `CI` environment variable.

This updates the comment to avoid that confusion.