net/server: close connections in parallel.
Previously we use ss::do_for_each to iterate residue connections
and close them while shutting down the server, which however is
not very suitable for this situation:

- ss::do_for_each emphasizes orderliness, which means one async
  operation won't be issued until the previous one resolves.
- ss::do_for_each may stop in the middle of the iteration, when
  the action is invoked and returns an exceptional future.

While closing the residue connections, we do not care about the
order since they are totally independent. Besides that, we also
want to close all the connections as much as possible.

Here we replace ss::do_for_each with ss::max_concurrent_for_each
to close all the connections in parallel, which can also ensure
the shutdown method will be called for every connection and limit
its concurrency to avoid reactor stall.

Signed-off-by: Jianyong Chen <balus@foxmail.com>