tests: extend trim_results to remove duplicate binaries
We have code in RedpandaService that saves a copy of the binary
when something crashes, but tests are not aware of each other's
binaries, so we can end up with duplicates in the tarball.

Hash them and remove the duplicates, to reduce the size
of the results tarball when multiple tests experience crashes.