cluster: add missing default equality operators
The envelope type has a default equality operator which means that
comparing types that inherit will still compare to each other, but use
the compartor in the base class (envelope). this is largely broken
because forgetting the operator== in the derived class means comparison
always succeeds. we are going to remove the operator in the base class
but these are still needed.

Note that the change in the commit to the creation of random timestamps
is important. Recently the treatment of timeout_clock (aka lowres_clock)
was changed to account for seastar not backing it by a milliseconds.

Those changes serialized the clock on the wire as milliseconds. However,
the new type backing lowres_clock is based on nanoseconds. Therefore
when scaling to milliseconds precision is lost. In many tests we were
generating durations that did not cleanly cast in both directions and
made equality tests very unreliable.

Signed-off-by: Noah Watkins <noah@redpanda.com>