kafka: simply the conditional operator expression
before this change, the return value of the assignment statement is
used as the right-hand operator of the assignment statement. this
is a valid use of these operators. but it would be more readable
if we could use use a plain value as the right-hand operator of
the assignment statement.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>