storage: capture more specific exception first
std::filesystem::filesystem_error is derived from std::system_error,
before this change, the latter is caught first, and then the former,
so std::filesystem::filesystem_error does not get a chance of being
caught despite that it is a more specific exception than
std::system_error. so we are not able to identify such an exception.

after this change, these two catch clauses are transposed, so we
can tell a `std::filesystem::filesystem_error` from a
`std::system_error`.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>