reflection/type_traits: use is_specialization_of template helper
use is_specialization_of<> template helper to check if a class template
is a specliazation of a container template classs.

this change is inspired by http://wg21.link/p2098. the only thing
we are missing is, as pointed by the proposal, the support for
templates like std::array<> and is_fragmented_vector<>.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>