raft: require full leadership before leadership transfer
Previously, we could have achieved an election win but
not yet have completed replication of configuration batches, and
be in a state where we have not learned the committed indices of
our followers.  This causes the needs_recovery check to fail,
and results in a timeout error that appears as a 500 to the
admin API user.

Instead, we should take the "I'm not the leader" path in this
situation, which prompts the admin API client with a 503 to
retry.

Fixes https://github.com/redpanda-data/redpanda/issues/5332

(cherry picked from commit 1b32994bc7484117a321638934d75d233305a1d3)