cloud_storage: Download manifest with aborted transations
Download tx manifest as part of the segment hydration in remote_segment.
Previously, the background hydration fiber in the remote_segment
downloaded the segment file and materialized the segment index. It
handled several cases: the segment might be already in the cache or not.
The index file might exist in cache or not.

This commit implements hydration of the aborted transactions metadata
(which is stored in tx manifest). This manifest has to be downloaded
alongside the segment. It's tiny compared to segment so normally it's
not a big deal. But the problem is that the number of cases that the
background fiber has to handle increases. Every file in cache can be
present or not present or be in progress. With two files we have 9
possible combinations.

The operation is now split into two steps:
1. Hydration (downloading) of data files in SI cache
2. Materialization of in-memory structures

The hydration step deals with complexity by eliminating impossible
combinations (in_progress state is not possible since we have only one
background fiber per remote_segment instance and only one remote_segment
instance per actual segment). It hydrates both tx manifest and segment
or only one of the files if another one is present.

The materialization step follows the hydration step. The segment file is
used to materialize the segment index and open a file handle. The tx
manifest file is used to build the list of aborted transactions (which
is stored in memory). If one of the materialization goals can't be met
(because the data file was evicted from cache) the background fiber will
just repeat the hydration step. When the materialization step is
completed the remote_segment is self contained and the files can be
freely evicted by cache.