c/types: introduce kafka offset types
We're going to mix raft and kafka offset in the same class, since
both the offsets uses the same type it's easy to make an error and
treat one as it was another. Introducing kafka offset to rely on the
type system to prevent such errors.

(cherry picked from commit 93350756123f6883192a0fb6bb2b123f394e9a4b)

Conflicts:
	src/v/cluster/types.h