Introduce type-erased any_handler
This is a polymorphic handler each of which is backed by an existing
concrete handler, but which lets us treat handlers generically without
restorting to template functions.

This reduces code bloat significantly as we do not duplicate code paths
for our ~45 handler types.

For example, requests.cc.o drops from ~11 MB to ~5 MB after it is
switched to the any_handler approach.