Introduce type-erased handler interface
This is a polymorphic handler class (abstract class with virtual
methods), as well as concrete instantiations of the interface for each
of the existing handlers, which can be looked up by API key.

This lets us treat handlers generically without resorting to template
functions which must be specialized for each handler type.

This reduces code bloat significantly as we do not duplicate code paths
for our ~45 handler types.

For example, requests.cc.o drops from ~11 MB to ~5 MB after it is
switched to the any_handler approach.