r/consensus: try updating leader commit index after aborting config change
When configuration change is aborted we append simple configuration
entry to each raft group member log. If node is a leader we must try to
update the committed index. For single replica raft groups that is the
only way to make the committed offset progress and finish
reconfiguration.

Updating committed offset is possible for single replica raft groups as
the node is the only source of truth for all raft decisions hence an
entry appended to leader log may immediately be committed.

Fixes: #5338

Signed-off-by: Michal Maslanka <michal@redpanda.com>