storage: don't save clean_segment on log close errors
This is the case where I/O errors occur trying to
close() out segments, and we risk a scenario
where something goes wrong flushing the segments
but we then succeed writing out a bogus clean_segment
record.

React to I/O errors by not marking the latest segment
clean at all, even if it wasn't the one that had
the error: the system is in a weird state and we
should exercise an abundance of caution.