s/compacted_index: prefix indexed key with batch type
Added batch type prefix to key stored in compaction index, both on disk
and in memory. This way a key stored in an index in actually a tuple
consisting of (batch_type, key_payload). Thanks to this approach
compaction logic is able to compact keys per batch type instead of
discriminating the batch type completely.

Signed-off-by: Michal Maslanka <michal@redpanda.com>