tests: rolling restart option for RedpandaService
This commit adds a new rolling_restart_nodes() method in RedpandaService
to orchestrate bits of a rolling restart. The actual bits to exercise
the rolling restart are placed in a separate file so we avoid ballooning
redpanda.py as our rolling restart capabilities and test requirements
expand.

It also adds a basic test to test this with a workload running in the
background.