cloud_storage: Propagate aborted transactions metadata
Add method that returns aborted txn's state in the remote_partition.
Also, make it available in the cluster partition.

The method hydrates segments in the requested range if needed. Normally
the hydration won't be needed because 'aborted_transactions' method is
called after the actual read from 'remote_partition' so the segments
will most likely be already hydrated and materialized.