c/topics_update_dispatcher: fixed de-allocation of deleted topics
Recently introduced two stage update of partition allocator changed the
way how we update allocation state. Currently when partition is being
moved allocations are added when processing update command and removed
when processing finished command.

When topic is being updated and it is deleted we must deallocate the
in-progress replicas that would normally be removed with the update
finsihed command.

Signed-off-by: Michal Maslanka <michal@redpanda.com>