random: add string generator with limited cardinality
This is a `generator` in the computer science sense; it generates a
sequence, like an iterator. The set of values this generator will output
is limited to a maximum size (cardinality). The distribution of values
is uniform. They are a fixed length, and the simple algorithm values
memory efficiency over actual entropy.

Useful for randomly generating keys that can be compacted.