s/test/util: basic log gap analysis for compaction tests
Introduce a utility for compaction testing which helps detect which
offset ranges were affected by compaction. This utility makes a single
pass over all of the message offsets in a given log instance,
and returning the total number of offset gaps, and the min and max
offsets that were missing.