cloud_storage: Remove assertion in access_time_tracker
Remove assertion from 'remote_timestamp' method of the
'access_time_tracker'. The original intention of the assertion was to
improve visibility in case of failure (if bad_alloc is thrown out of
erase method). This commit removes assertion and remplaces it with
logging. This won't cause any problems for cache other than leaking
some memory (since the element will remain in the collection). But it
will be removed eventually when cache eviction will call 'remove_others'
method. So it's not even a real memory leak.