config: increased default partitions # for __consumer_offsets
When a consumer tries to locate a consumer group coordinator of a cluster
for the first time, the __consumer_offsets topic is created with the
number of partitions as per the group_topic_partitions property.
The default value for that property was 1 which means that unless
a different value was explicitly specified by the customer at a very
early stage of cluster's life, all OffsetCommit requests from all
consumers will be going to a single broker. This change increases
the default value to 16 as a reasonable trade-off between OffsetCommit
parallelism for the clusters that will use consumer groups
later in their life, and the overhead for the clusters that
won't use consumer groups.

#5222