Use a conservative estimator for metadata requests
Currently we estimate that metadata requests take 8000 + rsize * 2 bytes
of memory to process, where rsize is the size of the request. Since
metadata requests are very small, this end up being roughly 8000 bytes.

However, metadata requests which return information about every
partition and replica may easily be several MBs in size.

To fix this for metadata requests specifically, we use a new more
conservative estimate which uses the "maximum supported"
partition count give an upper bound on the size. This this ends up
"only" 8 MB, and applies only to metadata requests, the performance
cost (chiefly a reduced maximum number of metadata requests in flight on
one shard) should be moderate.