ducktape: run GroupMetricsTest with group_topic_partitions=1
test_leadership_transfer relies on kafka_group_offset metric source as an
indication of the node being a group coordinator. However in admin API
there is no way to know which __consumer_groups partition a specific
consumer group uses, and the testcase always works with partition 0.
To make that work, number of __consumer_groups partition is set to 1
explicitly.