tests: re-use installs in upgrade tests
Tests that use the RedpandaInstaller are bandwidth-intensive and often
take several minutes to complete on account of the download of hundreds
of MBs worth of tarballs.

This commit mitigates this in local ducktape by having all test
containers share a single bind mount. The installer now uses a lock file
to prevent concurrent operations on the mount (e.g. when downloading
binaries, checking to see what binaries exist, etc).

With this commit, regardless of whether in local or clustered ducktape,
we also no longer get rid of downloaded binaries between test runs.
Instead, after a test completes, we just revert any changes to the
original binaries, and leave the rest be.