c/feature_table: align logical versions with v22.1.5
We use feature manager to postpone activating a feature until all nodes
in the cluster are updated to avoid a situation in which a node with a
newer version uses an updated version of a protocol to communicate with
an older version which doesn't understand it.

We tend to avoid backporting features but there are cases when we need
to do it anyway e.g. when a bug fix introduces a protocol change and we
need to use the feature manager to perform a non-interruptive upgrade.

When we do it we need to do it very carefully to preserve the feature
manager invariant:

   the series of versions and the definitions of the features must
   line up perfectly with the history of versions+features in subsequent
   releases

   https://vectorizedio.atlassian.net/wiki/spaces/CORE/pages/170491930

However if we the bugfix is added after we already introduced several
feature which we don't backport we need to update the dev branch (future
release) to be consistent with the minor "backport" release.

This change updates the dev to be consistent with v22.1.5 release which
contains a bug fix gated by rm_stm_kafka_cache feature flag.