rpc: add new service methods for testing
Adds three new test service methods:

  - echo_adl_only
  - echo_serde_only
  - echo_adl_serde

Each of these methods has normal echo behavior, but uses separate types
for each method specific to adl and serde support. For example,
echo_adl_serde uses types that are both adl and serde encodable.

Furthermore, each encoding/decoding procedure modifies the content of
the message being echoed with an indication of the operation performed.
For example, encoding a message with adl will result in a suffix:

    <orig>_to_aao

where in _to_[a][ao] the first [a] indicates adl encoding and the [ao]
indicates an adl-only type. _to and _from are encoding and decoding,
respectively. So a round-trip RPC echo might result in a final result
arriving at the client with the string:

    <orig>_to_aao_from_aao_to_aao_from_aao

This simple mechanism allows writing tests that verify low-level
behavior regarding how different types are handled without resorting to
having to track and export stats.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>