rpc: transport request adaptive encoding selection
When a message is sent the active transport version determines if it is
sent with adl (v0,1) or serde encoding (v2). Serde-only messages require
a version >= v2. Messages with adl-only support are always sent at v0.
Support for adl-only messages is a temporary solution to allow changes
to the RPC transport and servers to be worked on prior to all types
being supported by serde. Once all types have serde support the
allowance at the type-system level should be removed.  Sending a message
does not upgrade the transport. That is done when handling the reply.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>