rptest: test against the arroyo test suite
This commit adds a ducktape test that runs the arroyo test suite
against a redpanda cluster. All test output is logged and test failures
are parsed from the pytest output. The test fails if there were pytest
failures or if pytest returns an error code indicating that the tests
couldn't run.