Add "unhealthy reason" to health monitor
Currently, the health monitor returns a boolean indicating if the
cluster is healty, and some additional information about nodes which
are down and leaderless partitions.

However, it is possible for the "healthy" boolean to be false while
the remainder of the information does not indicate any problem,
leaving users of rpk cluster health with little to go on: "The cluster
is unhealthy, but you don't get to know why".

One way (only way?) this can happen is when the cluster health
report acquisition process fails.

This change adds an unhealthy_reason string, which is set to non-empty
only if the cluster is unhealthy, and which contains one reason for
that state. For example if the report couldn't be obtained, it gives
details about that error.