rpk redpanda config set: improve setting arrays
Recent rpk improvements now initialize index 0 of a slice if the key we
are setting is deeply inside the slice. e.g., foo.bar.baz would
initialize bar[0] to set baz. However, it was not possible to set
foo.bar[1]. This change fixes that. We also allow setting one past the
end of an array to extend it: if foo.bar[0] exists, then foo.bar[1].baz
adds to the array.

Old rpk never had the ability to set an array key to a single value that
was not wrapped in brackets. For example, `foo.bar = "baz"` would not
work if bar was an array. Now, we use / initialize element 0 and set
that.

Fixes #2958
Fixes #5498
Fixes #5264