rm_stm: use last stable offset to limit compaction
Last Stable Offset (LSO) is the largest offset such that no messages, up
to and including LSO, are part of an unresolved transaction.

Use LSO to determine the maximum offset that may be affected by
compaction. This ensures that messages for open transactions--which may
be aborted or committed in the future--cannot affect compaction. For
example, an aborted message cannot replace a previous message with the
same key, during compaction.