rpc: set reply version on error
When an error occurs while processing a request (e.g. a handler throws
or the server receives an unsupported api request) a default initialized
netbuf is created to hold the error and is sent back to the client.

Prior to this commit the header included v0 as the version for such
error replies. Since the error may be returned to a client whose
transport was upgraded to v2 a protocol violation error would be logged
by the client.

This commit matches the reply version to the request version for error
replies under the reasonable assumption that (1) the rpc header is
compatibile across transport versions and (2) the errors contain no
decodable payloads.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>