omb/perf: Bump client request timeouts to 5mins
Current timeout of 2mins does not seem sufficient in high throughput
runs (> 1GBps) and we run into this client side timeout issue and the
test is flagged as a failure.

java.util.concurrent.CompletionException: \
org.apache.kafka.common.errors.TimeoutException: \
Expiring xxx record(s) for <topic>: 1200xx ms has passed since batch creation

This exception means that the records are lingering on the client side
longer than they should. While this can happen for a variety of reasons,
typically under high load this means that the broker is not able to keep
up with the produce rate. Typical way to solving this

* Add more brokers (load spread)
* Reduce linger.ms (batches don't linger for long)
* Reduce batch.size (batches are shipped more often)
* Increase request timeout.

We are benchmarking on a 3 node cluster and its probably not prudent to
add more nodes in a nightly run ($$$). We run with 1ms linger.ms, which
is already quite low and the batch size is decently low. So we take up
the last option of increasing timeout.

This could potentially mean that we might see latency spikes (and high
P9x) if we let this batches go through, but that is okay. There is an
orthogonal effort to chase down high P9xs and we fix it as a part of
that effort.

I arrived at this 5minute timeout empirically by doing a few runs on the
same cluster spec in ec2 and the perf run seems to pass normally as
expected.