r/group_configuration: introducing change_strategy
Introduced `change_strategy` type that is going to encapsulate logic
related with changing raft group configuration for different raft group
configuration versions.

For `raft::group_configuration` changing version may involve not only
change in serialized format but also change in the configuration
update/state transition logic. Having the change strategy abstracted we
may change the logic of configuration change in runtime based on
deserialized configuration version.

Virtual dispatch overhead is not the problem for configuration changes
as they are not in the hot path and not very often.

Signed-off-by: Michal Maslanka <michal@redpanda.com>